requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pillow==11.1.0
pydantic==2.10.4
httpx==0.28.1
tenacity==9.0.0
